% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundNormalizedVariable.R
\name{BoundNormalizedVariable}
\alias{BoundNormalizedVariable}
\title{Take a normalized variable and transform it back to a bounded variable.}
\usage{
BoundNormalizedVariable(x, constraints)
}
\arguments{
\item{x}{A vector, matrix, array, or dataframe with value to be coerced into a range or set.}

\item{constraints}{A named list of constraints with optional \code{lower} and/or
\code{upper} elements. See the examples below for formatting details.}
}
\value{
An object of the same class as x with the values transformed into the desired half-line or segment.
}
\description{
This takes variables on the real line and constrains them to be on
a half-line (constrained above or below) or a segment (constrained both
above and below). This is approximately the inverse of
\code{NormalizeBoundedVariable}; this does not completely reverse the
effect of \code{NormalizeBoundedVariable} because \code{NormalizeBoundedVariable}
first forces values away from the bounds, and this information is lost.
No \code{tol} parameter is needed here because the inverse transformations
(\code{exp}, \code{pnorm}) naturally keep output away from the boundaries
for any finite input.
}
\examples{
  BoundNormalizedVariable(c(-2, 0, 2), constraints = list(lower = 0))
  BoundNormalizedVariable(c(-2, 0, 2), constraints = list(upper = 5))
  BoundNormalizedVariable(c(-2, 0, 2), constraints = list(lower = 0, upper = 5))
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
