% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_imputation.R
\name{fast_imputation}
\alias{fast_imputation}
\title{Use patterns learned from training data to impute missing values}
\usage{
fast_imputation(x, patterns, verbose = FALSE)
}
\arguments{
\item{x}{A data frame or tibble, possibly with some missing (\code{NA})
values. Must have the same columns in the same order as the training data
passed to \code{\link{train_fast_imputation}}.}

\item{patterns}{An object of class \code{fast_imputation_patterns} generated
by \code{\link{train_fast_imputation}}.}

\item{verbose}{If \code{TRUE} a progress bar is shown. Default \code{FALSE}.}
}
\value{
\code{x} with missing values filled in. Returns a tibble if
  \code{x} is a tibble, otherwise a data frame.
}
\description{
Like Amelia, FastImputation assumes that the columns of the data are
multivariate normal or can be transformed into approximately
multivariate normal.
}
\examples{
data(FI_train)
patterns <- train_fast_imputation(
  FI_train,
  constraints = list(
    bounded_below_2           = list(lower = 0),
    bounded_above_and_below_6 = list(lower = 0, upper = 1)
  ),
  ignore_cols = user_id_1,
  categorical = categorical_9
)

data(FI_test)
imputed_data <- fast_imputation(FI_test, patterns)

}
\references{
\url{https://gking.harvard.edu/amelia}
}
\seealso{
\code{\link{train_fast_imputation}}
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
