% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastImputation.R
\name{FastImputation}
\alias{FastImputation}
\title{Use the pattern learned from the training data to impute (fill in good guesses for) missing values.}
\usage{
FastImputation(x, patterns, verbose = TRUE)
}
\arguments{
\item{x}{Vector, matrix, dataframe, or object that can be coerced into a dataframe, possibly with some missing (\code{NA}) values.}

\item{patterns}{An object of class 'FastImputationPatterns' generated by \code{TrainFastImputation}.}

\item{verbose}{If TRUE then the progress in imputing the data will be shown.}
}
\value{
x, but with missing values filled in (imputed)
}
\description{
Like Amelia, FastImputation assumes that the columns of the data are
multivariate normal or can be transformed into approximately
multivariate normal.
}
\examples{
data(FI_train)   # provides FItrain dataset
patterns <- TrainFastImputation(
  FI_train,
  constraints=list(list(1, list(ignore=TRUE)),       # use for ids, other vars not to impute
                   list(2, list(lower=0)),           # continuous var with lower bound
                   list(5, list(upper=0)),           # continuous var with only upper bound
                   list(6, list(lower=0, upper=1)),  # bounded to a finite interval
                   list(9, list(categorical=TRUE))   # finite number of values
                   ))                                # vars 3, 4, 7, 8 continuous with no bounds

data(FI_test)
FI_test          # note there is missing data
\dontrun{imputed.data <- FastImputation(FI_test, patterns)}
\dontrun{imputed.data    # good guesses for missing values are filled in}

data(FI_true)
\dontrun{imputation.rmse <- sqrt(sum( (imputed.data - FI_true)^2 )/sum(is.na(FI_test)))}
\dontrun{imputation.rmse}
}
\references{
\url{http://gking.harvard.edu/amelia/}
}
\seealso{
\code{\link{TrainFastImputation}}
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
