% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fast_imputation_patterns}
\alias{print.fast_imputation_patterns}
\title{Print a fast_imputation_patterns object}
\usage{
\method{print}{fast_imputation_patterns}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fast_imputation_patterns}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fast_imputation_patterns object
}
