% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_fast_imputation.R
\name{train_fast_imputation}
\alias{train_fast_imputation}
\title{Learn from training data so that later you can fill in missing data}
\usage{
train_fast_imputation(
  x,
  constraints = list(),
  ignore_cols = NULL,
  categorical = NULL
)
}
\arguments{
\item{x}{A data frame (or tibble) containing training data. May have
incomplete rows.}

\item{constraints}{A named list of bound constraints. Each element is named
after a column in \code{x} and is itself a list with optional
\code{lower} and/or \code{upper} numeric elements.}

\item{ignore_cols}{<[`tidy-select`][tidyselect::language]> Columns to
exclude from imputation (e.g. ID columns). These columns are preserved
in the output unchanged.}

\item{categorical}{<[`tidy-select`][tidyselect::language]> Columns that
contain a finite set of discrete values (factors/characters). These are
one-hot encoded internally.}
}
\value{
An object of class \code{fast_imputation_patterns} that contains
  all information needed to impute missing values with
  \code{\link{fast_imputation}}.
}
\description{
Like Amelia, FastImputation assumes that the columns of the data are
multivariate normal or can be transformed into approximately
multivariate normal.
}
\examples{
data(FI_train)

patterns <- train_fast_imputation(
  FI_train,
  constraints = list(
    bounded_below_2           = list(lower = 0),
    bounded_above_5           = list(upper = 0),
    bounded_above_and_below_6 = list(lower = 0, upper = 1)
  ),
  ignore_cols = user_id_1,
  categorical = categorical_9
)

}
\references{
\url{https://gking.harvard.edu/amelia}
}
\seealso{
\code{\link{fast_imputation}}
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
\keyword{internal}
