% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainFastImputation.R
\name{TrainFastImputation}
\alias{TrainFastImputation}
\title{Learn from the training data so that later you can fill in missing data}
\usage{
TrainFastImputation(x, constraints = list())
}
\arguments{
\item{x}{Dataframe containing training data. Can have incomplete rows.}

\item{constraints}{A list of constraints.  See the examples below for formatting details.}
}
\value{
An object of class 'FastImputationPatterns' that contains
  information needed later to impute on a single row.
}
\description{
Like Amelia, FastImputation assumes that the columns of the data are
multivariate normal or can be transformed into approximately
multivariate normal.
}
\examples{

data(FI_train)   # provides FI_train dataset

patterns_with_constraints <- TrainFastImputation(
  FI_train,
  constraints=list(list(1, list(ignore=TRUE)),       # use for ids, other vars not to impute
                   list(2, list(lower=0)),           # continuous var with lower bound
                   list(5, list(upper=0)),           # continuous var with only upper bound
                   list(6, list(lower=0, upper=1)),  # bounded to a finite interval
                   list(9, list(categorical=TRUE))   # finite number of values
                   ))                                # vars 3, 4, 7, 8 continuous with no bounds
}
\references{
\url{http://gking.harvard.edu/amelia/}
}
\seealso{
\code{\link{FastImputation}}
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
