% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichAreFactors.R
\name{WhichAreFactors}
\alias{WhichAreFactors}
\title{Return the names of columns in a data.frame that are factors.}
\usage{
WhichAreFactors(x)
}
\arguments{
\item{x}{A data.frame to check for factor columns.}
}
\value{
A character vector of column names that are factors.
}
\description{
Return the names of columns in a data.frame that are factors.
}
