% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainFastImputation.R
\name{TrainFastImputation}
\alias{TrainFastImputation}
\title{Learn from the training data so that later you can fill in missing data}
\usage{
TrainFastImputation(x, constraints, ignore_cols, categorical)
}
\arguments{
\item{x}{Dataframe containing training data. Can have incomplete rows.}

\item{constraints}{A named list of constraints.  See the examples below for formatting details.}

\item{ignore_cols}{A vector of column indices or column names to be ignored in the imputation process.}

\item{categorical}{A vector of column indices or column names of varaibles with a finite set of possible values.}
}
\value{
An object of class 'FastImputationPatterns' that contains
  information needed later to impute on a single row.
}
\description{
Like Amelia, FastImputation assumes that the columns of the data are
multivariate normal or can be transformed into approximately
multivariate normal.
}
\examples{
# Small fast example
train_df <- data.frame(x = c(1, 2, NA, 4, 5), y = c(2, NA, 6, 8, 10))
TrainFastImputation(train_df)

\donttest{
# Full example with package dataset
data(FI_train)
patterns_with_constraints <- TrainFastImputation(
  FI_train,
  constraints=list("bounded_below_2" = list(lower=0),
                   "bounded_above_5" = list(upper=0),
                   "bounded_above_and_below_6" = list(lower=0, upper=1)
                   ),
  ignore_cols="user_id_1",
  categorical="categorical_9")
}

}
\references{
\url{https://gking.harvard.edu/amelia}
}
\seealso{
\code{\link{FastImputation}}
}
\author{
Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}
