\name{UnfactorColumns}
\alias{UnfactorColumns}
\title{Convert columns of a dataframe from factors to character or numeric.}
\usage{
  UnfactorColumns(x)
}
\arguments{
  \item{x}{A dataframe}
}
\value{
  A dataframe containing the same data but any
  \code{factor} columns have been replaced with numeric or
  character columns.
}
\description{
  Convert columns of a dataframe from factors to character
  or numeric.
}
\author{
  Stephen R. Haptonstahl \email{srh@haptonstahl.org}
}

